Testing_task/photoDuplicate/requirements.txt
numpy~=1.26.4
pillow~=10.3.0
ImageHash~=4.3.1